+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Fixed bug in `Determinize`: `HashMap` -> `BTreeMap` for `LabelMap`.
+- `NO_LABEL` and `NO_STATE_ID` are now public. 
 - Fixed bug in `RmEpsilon`: mutation had to be performed while iterating.
 - `QuantizeMapper` has now a configurable delta value with default set to `KDELTA`
 - Replace `FstCache` implementation for `Arc<FstCache>` by the deref trait.